Add go.mod
This adds a go.mod to this project; note that (given that this project
already reached `v2.0.0`) the _correct_ module name would be
`github.com/mistifyio/go-zfs/v2`. As a result, projects using this
module will see a `+incompat` comment for this module in their go.mod.

Changing the module name to `v2` in itself would be a breaking change
however, so would require an update to the _major_ version (to follow
SemVer), and thus require `v3.0.0` to be released.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>